Fixed compiler warning (love u gemini) and added an easter egg (hint: 69 777)

+++ code_files/processes.cpp
@@ -7,7 +7,8 @@ Processes::~Processes() {
 }
 
 void Processes::join_threads() {
-    for (int i = 0; i < this->threads.size(); i++) {
+    // Use size_t for the loop counter to match the type of std::vector::size(), avoiding a compiler warning.
+    for (size_t i = 0; i < this->threads.size(); i++) {
         if (this->threads[i] && this->threads[i]->joinable()) {
             this->threads[i]->join(); // Use -> to call join() on the pointer
         }
@@ -15,7 +16,8 @@ void Processes::join_threads() {
 }
 
 void Processes::clear_threads() {
-    for (int i = 0; i < this->threads.size(); i++) {
+    // Use size_t for the loop counter to match the type of std::vector::size(), avoiding a compiler warning.
+    for (size_t i = 0; i < this->threads.size(); i++) {
         delete this->threads[i]; // Free the dynamically allocated memory
     }
     this->threads.clear();
@@ -26,7 +28,9 @@ void Processes::add_thread(std::thread* thread) {
 }
 
 void Processes::remove_thread(int index) {
-    if (index >= 0 && index < this->threads.size()) {
+    // Check if the index is non-negative and within the bounds of the vector.
+    // The index is cast to size_t to safely compare with this->threads.size() and avoid a signed/unsigned mismatch warning.
+    if (index >= 0 && static_cast<size_t>(index) < this->threads.size()) {
         if (this->threads[index] && this->threads[index]->joinable()) {
             this->threads[index]->join(); // Ensure the thread is joined before removing
         }
